+++ tests/meltano/core/sql/test_query_generation.py
@@ -39,47 +39,100 @@ def setup_test_models(project, request):
 
 
 @pytest.fixture(scope="class")
-def design_helper(project, setup_test_models):
+def gitflix(project, setup_test_models):
     # Compile the test files and make sure that the proper m5oc file was generated
     ProjectCompiler(project).compile()
 
     # Load the m5oc file for gitflix
     m5oc_file = project.root_dir("model", "gitflix.topic.m5oc")
-    m5oc = M5ocFile.load(m5oc_file)
-    return m5oc.design("users")
+    return M5ocFile.load(m5oc_file)
 
 
-@pytest.fixture(scope="module")
-def query_payload():
-    payload = """
-    {
-     "run":true,
-     "table":"users",
-     "columns":["gender"],
-     "aggregates":["count", "avg_age", "sum_clv"],
-     "timeframes":[],
-     "joins":[
-      {"name":"streams","columns":["day", "month", "year"],"aggregates":["count", "sum_minutes", "count_days"]},
-      {"name":"episodes","columns":["tv_series"],"aggregates":["count", "avg_rating"]}
-     ],
-     "order":null,
-     "limit":"50",
-     "filters":{}
-    }
-    """
-    return json.loads(payload)
+class PayloadBuilder:
+    def __init__(self, table: str):
+        self._table = table
+        self._columns = set()
+        self._aggregates = set()
+        self._joins = {}
+
+    def join(self, name: str):
+        if name not in self._joins:
+            self._joins[name] = PayloadBuilder(name)
+
+        return self._joins[name]
+
+    def columns(self, *columns, join=None):
+        if join:
+            self.join(join).columns(*columns)
+        else:
+            self._columns.update(columns)
+
+        return self
+
+    def aggregates(self, *aggregates, join=None):
+        if join:
+            self.join(join).aggregates(*aggregates)
+        else:
+            self._aggregates.update(aggregates)
+
+        return self
+
+    def as_join(self):
+        return {
+            "name": self._table,
+            "columns": self._columns,
+            "aggregates": self._aggregates,
+        }
+
+    @property
+    def payload(self):
+        return {
+            "run": True,
+            "table": self._table,
+            "columns": self._columns,
+            "aggregates": self._aggregates,
+            "timeframes": [],
+            "joins": [join.as_join() for join in self._joins.values()],
+            "order": None,
+            "limit": "50",
+            "filters": {},
+        }
 
 
 class TestQueryGeneration:
-    def test_compile_and_load_m5o_files(self, project, design_helper):
-        design = MeltanoDesign(definition=design_helper.design)
+    @pytest.fixture
+    def users(self):
+        return (
+            PayloadBuilder("users")
+            .columns("gender")
+            .aggregates("count", "avg_age", "sum_clv")
+            .columns("day", "month", "year", join="streams")
+            .aggregates("count", "sum_minutes", "count_days", join="streams")
+            .columns("tv_series", join="episodes")
+            .aggregates("count", "avg_rating", join="episodes")
+        )
+
+    @pytest.fixture
+    def streams(self):
+        return (
+            PayloadBuilder("streams")
+            .columns("day", "month", "year")
+            .aggregates("count", "sum_minutes", "count_days")
+            .columns("gender", join="users")
+            .aggregates("count", "avg_age", "sum_clv", join="users")
+            .columns("tv_series", join="episodes")
+            .aggregates("count", "avg_rating", join="episodes")
+        )
+
+    def test_compile_and_load_m5o_files(self, project, gitflix):
+        design = MeltanoDesign(definition=gitflix.design("users").design)
 
         assert design.name == "users"
         assert len(design.tables()) == 3
         assert len(design.joins()) == 2
 
-    def test_meltano_base_classes(self, design_helper):
-        design = MeltanoDesign(definition=design_helper.design)
+    def test_meltano_base_classes(self, gitflix):
+        design = MeltanoDesign(definition=gitflix.design("users").design)
 
         assert design.name == "users"
         assert len(design.tables()) == 3
@@ -119,10 +172,12 @@ class TestQueryGeneration:
         table.add_aggregate(new_a)
         assert len(table.aggregates()) == 4
 
-    def test_meltano_query(self, design_helper, query_payload):
+    def test_meltano_query(self, users, gitflix):
         # Test parsing a json payload using a Design generated from a m5oc file
         #  and generating a proper MeltanoQuery Object
-        q = MeltanoQuery(definition=query_payload, design_helper=design_helper)
+        q = MeltanoQuery(
+            definition=users.payload, design_helper=gitflix.design("users")
+        )
 
         assert q.design.name == "users"
         assert len(q.tables) == 3
@@ -142,6 +197,43 @@ class TestQueryGeneration:
         assert 'COALESCE(COUNT("users.id"),0)' in sql
         assert 'COALESCE(SUM("users.clv"),0)' in sql
         assert 'SELECT * FROM "result"' in sql
+        assert 'JOIN "streams"' in sql
+        assert 'JOIN "episodes"' in sql
+
+        # Check that the attribute used both as a column and as an aggregate
+        # 1. Only appears once in the select clause of the base query
+        # 2. Properly appears as an aggregate
+        assert sql.count('"streams"."day" "streams.day"') == 1
+        assert 'COALESCE(COUNT("streams.day"),0)' in sql
+
+    def test_meltano_disjointed_query(self, streams, gitflix):
+        # Test parsing a json payload using a Design generated from a m5oc file
+        #  and generating a proper MeltanoQuery Object
+        q = MeltanoQuery(
+            definition=streams.payload, design_helper=gitflix.design("streams")
+        )
+
+        assert q.design.name == "streams"
+        assert len(q.tables) == 3
+        assert len(q.join_order) == 3
+        assert q.join_order[2]["table"] == "episodes"
+
+        # Test generating an HDA query
+        (sql, column_headers, column_names, aggregate_columns) = q.hda_query()
+
+        assert "Average Age" in column_headers
+        assert "sum_minutes" in column_names
+        assert "users.sum_clv" in aggregate_columns
+
+        assert "WITH base_join AS (SELECT" in sql
+        assert "base_streams AS (SELECT DISTINCT" in sql
+        assert "users_stats AS (" in sql
+        assert 'COALESCE(AVG("episodes.rating"),0)' in sql
+        assert 'COALESCE(COUNT("users.id"),0)' in sql
+        assert 'COALESCE(SUM("users.clv"),0)' in sql
+        assert 'SELECT * FROM "result"' in sql
+        assert 'JOIN "users"' in sql
+        assert 'JOIN "episodes"' in sql
 
         # Check that the attribute used both as a column and as an aggregate
         # 1. Only appears once in the select clause of the base query