+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Added documentation on custom transformations and models
 
 ### Fixes
 * [#555](https://gitlab.com/meltano/meltano/issues/555) Fix `meltano discover` improperly displaying plugins
+* [#530](https://gitlab.com/meltano/meltano/issues/530) Fix query generation for star schemas
 
 ### Breaks